Extract colored division edges into a ColoredDivisionEdges helper
The division-edge mechanism (~180 lines: connection-row reconstruction, in-place
data augmentation, display_graph suppression, restore) lived inside
SpacetimeWidget. Move it to division_edges.ColoredDivisionEdges, a self-contained
unit the widget owns and drives around the lift: apply(layers) before
engine.apply, finalize(viewer) after, teardown() on toggle-off/recompute/dispose.

The widget keeps only the policy -- checkbox gate, which role layers, and the
user-facing warnings -- in a slim _rebuild_division_edges. _widget.py drops from
879 to 700 lines (numpy import no longer needed there); the mechanism is now
testable on its own. Behaviour unchanged: verified build/teardown end-to-end and
the full non-GUI suite (86 passed).
